@@ -1,71 +1,69 @@
 {
-    "name": "ionic2-weather",
-    "version": "0.0.1",
-    "author": "Ionic Framework",
-    "homepage": "http://ionicframework.com/",
-    "private": true,
-    "scripts": {
-        "clean": "ionic-app-scripts clean",
-        "build": "ionic-app-scripts build",
-        "lint": "ionic-app-scripts lint",
-        "ionic:build": "ionic-app-scripts build",
-        "ionic:serve": "ionic-app-scripts serve"
+  "name": "ionic2-weather",
+  "version": "0.0.1",
+  "author": "Ionic Framework",
+  "homepage": "http://ionicframework.com/",
+  "private": true,
+  "scripts": {
+    "clean": "ionic-app-scripts clean",
+    "build": "ionic-app-scripts build",
+    "lint": "ionic-app-scripts lint",
+    "ionic:build": "ionic-app-scripts build",
+    "ionic:serve": "ionic-app-scripts serve"
+  },
+  "dependencies": {
+    "@angular/common": "5.0.0",
+    "@angular/compiler": "5.0.0",
+    "@angular/compiler-cli": "5.0.0",
+    "@angular/core": "5.0.0",
+    "@angular/forms": "5.0.0",
+    "@angular/http": "5.0.0",
+    "@angular/platform-browser": "5.0.0",
+    "@angular/platform-browser-dynamic": "5.0.0",
+    "@ionic-native/core": "4.3.3",
+    "@ionic-native/geolocation": "4.3.3",
+    "@ionic-native/splash-screen": "4.3.3",
+    "@ionic-native/status-bar": "4.3.3",
+    "@ionic/storage": "2.1.3",
+    "cordova-android": "^6.2.3",
+    "cordova-plugin-compat": "^1.1.0",
+    "cordova-plugin-console": "^1.0.5",
+    "cordova-plugin-device": "^1.1.4",
+    "cordova-plugin-geolocation": "^2.4.3",
+    "cordova-plugin-splashscreen": "^4.0.3",
+    "cordova-plugin-statusbar": "^2.2.2",
+    "cordova-plugin-whitelist": "^1.3.1",
+    "ionic-angular": "3.9.2",
+    "ionic-plugin-keyboard": "^2.2.1",
+    "ionicons": "3.0.0",
+    "moment": "^2.18.1",
+    "rxjs": "5.5.2",
+    "sw-toolbox": "3.6.0",
+    "zone.js": "0.8.18"
+  },
+  "devDependencies": {
+    "@ionic/app-scripts": "3.1.0",
+    "dotenv": "^4.0.0",
+    "ionic": "3.18.0",
+    "lodash": "^4.17.4",
+    "typescript": "2.5.3"
+  },
+  "description": "An Ionic project",
+  "cordova": {
+    "plugins": {
+      "cordova-plugin-geolocation": {},
+      "cordova-plugin-whitelist": {},
+      "cordova-plugin-console": {},
+      "cordova-plugin-device": {},
+      "cordova-plugin-splashscreen": {},
+      "cordova-plugin-statusbar": {},
+      "ionic-plugin-keyboard": {}
     },
-    "dependencies": {
-        "@angular/common": "4.4.4",
-        "@angular/compiler": "4.4.4",
-        "@angular/compiler-cli": "4.4.4",
-        "@angular/core": "4.4.4",
-        "@angular/forms": "4.4.4",
-        "@angular/http": "4.4.4",
-        "@angular/platform-browser": "4.4.4",
-        "@angular/platform-browser-dynamic": "4.4.4",
-        "@ionic-native/core": "3.12.1",
-        "@ionic-native/geolocation": "^3.12.1",
-        "@ionic-native/splash-screen": "3.12.1",
-        "@ionic-native/status-bar": "3.12.1",
-        "@ionic/app-scripts": "2.1.3",
-        "@ionic/storage": "2.0.1",
-        "cordova-android": "^6.2.3",
-        "cordova-plugin-compat": "^1.1.0",
-        "cordova-plugin-console": "^1.0.5",
-        "cordova-plugin-device": "^1.1.4",
-        "cordova-plugin-geolocation": "^2.4.3",
-        "cordova-plugin-splashscreen": "^4.0.3",
-        "cordova-plugin-statusbar": "^2.2.2",
-        "cordova-plugin-whitelist": "^1.3.1",
-        "ionic-angular": "3.8.0",
-        "ionic-plugin-keyboard": "^2.2.1",
-        "ionicons": "3.0.0",
-        "moment": "^2.18.1",
-        "rxjs": "5.4.0",
-        "skycons": "^1.0.0",
-        "sw-toolbox": "3.6.0",
-        "zone.js": "0.8.18"
-    },
-    "devDependencies": {
-        "@ionic/app-scripts": "^3.0.1",
-        "dotenv": "^4.0.0",
-        "ionic": "3.16.0",
-        "lodash": "^4.17.4",
-        "typescript": "2.3.4"
-    },
-    "description": "An Ionic project",
-    "cordova": {
-        "plugins": {
-            "cordova-plugin-geolocation": {},
-            "cordova-plugin-whitelist": {},
-            "cordova-plugin-console": {},
-            "cordova-plugin-device": {},
-            "cordova-plugin-splashscreen": {},
-            "cordova-plugin-statusbar": {},
-            "ionic-plugin-keyboard": {}
-        },
-        "platforms": [
-            "android"
-        ]
-    },
-    "config": {
-        "ionic_webpack": "./config/webpack.config.js"
-    }
+    "platforms": [
+      "android"
+    ]
+  },
+  "config": {
+    "ionic_webpack": "./config/webpack.config.js"
+  }
 }
