@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { Http } from '@angular/http';
-import { Observable } from 'rxjs/Rx'
+import { Observable } from 'rxjs/Rx';
 import 'rxjs/add/operator/map';
 
 import { Forecast } from '../interfaces/forecast';
@@ -9,7 +9,7 @@ import { AppConfig } from '../config/app.config';
 @Injectable()
 export class ForecastServiceProvider {
   apiUrl = '/api/forecast';
-  
+
   constructor(public http: Http, private appConfig: AppConfig) {}
 
   load(latitude: number, longtude: number): Observable<Forecast> {
